# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IdeaGPT API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/roadmaps": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Create Roadmap",
        "operationId": "create_roadmap_api_v1_projects__project_id__roadmaps_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Get Project Roadmaps",
        "operationId": "get_project_roadmaps_api_v1_projects__project_id__roadmaps_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoadmapResponse"
                  },
                  "title": "Response Get Project Roadmaps Api V1 Projects  Project Id  Roadmaps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roadmaps/{roadmap_id}": {
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Get Roadmap",
        "operationId": "get_roadmap_api_v1_roadmaps__roadmap_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Update Roadmap",
        "operationId": "update_roadmap_api_v1_roadmaps__roadmap_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Delete Roadmap",
        "operationId": "delete_roadmap_api_v1_roadmaps__roadmap_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roadmaps/{roadmap_id}/tasks": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Add custom milestone task to roadmap",
        "operationId": "add_custom_task_api_v1_roadmaps__roadmap_id__tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roadmaps/{roadmap_id}/tasks/{task_id}/status": {
      "put": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Update roadmap task status (PENDING | IN_PROGRESS | COMPLETED)",
        "operationId": "update_task_status_api_v1_roadmaps__roadmap_id__tasks__task_id__status_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roadmaps/{roadmap_id}/critical-path": {
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Analyze roadmap dependency graph and compute critical chain",
        "operationId": "get_critical_path_api_v1_roadmaps__roadmap_id__critical_path_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomTaskRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "phase": {
            "type": "string",
            "maxLength": 50,
            "title": "Phase",
            "default": "MVP"
          },
          "priority": {
            "type": "string",
            "maxLength": 20,
            "title": "Priority",
            "default": "HIGH"
          },
          "estimated_days": {
            "type": "integer",
            "maximum": 180.0,
            "minimum": 1.0,
            "title": "Estimated Days",
            "default": 3
          },
          "dependencies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencies"
          },
          "milestone_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CustomTaskRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MilestoneSchema": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the milestone"
          },
          "objective": {
            "type": "string",
            "title": "Objective",
            "description": "Objective of this milestone"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSchema"
            },
            "type": "array",
            "title": "Tasks",
            "description": "List of tasks in this milestone"
          }
        },
        "type": "object",
        "required": [
          "title",
          "objective"
        ],
        "title": "MilestoneSchema"
      },
...
      "RoadmapCreate": {
        "properties": {
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneSchema"
            },
            "type": "array",
            "title": "Milestones",
            "description": "Validated JSON structure for roadmap milestones"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "title": "RoadmapCreate"
      },
...
      "RoadmapResponse": {
        "properties": {
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneSchema"
            },
            "type": "array",
            "title": "Milestones",
            "description": "Validated JSON structure for roadmap milestones"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "created_at",
          "updated_at"
        ],
        "title": "RoadmapResponse"
      },
      "RoadmapUpdate": {
        "properties": {
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneSchema"
            },
            "type": "array",
            "title": "Milestones",
            "description": "Validated JSON structure for roadmap milestones"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "title": "RoadmapUpdate"
      },
...
      "TaskSchema": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the task"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Detailed description of the task"
          },
          "estimated_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Days",
            "description": "Estimated effort in days"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status: pending, in_progress, completed",
            "default": "pending"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TaskSchema"
      },
      "TaskStatusUpdateRequest": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(PENDING|IN_PROGRESS|COMPLETED)$",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TaskStatusUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}